backend/requirements.txt
fastapi
pydantic
requests
uvicorn
openai>=1.0.0
python-dotenv
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
